Make `SearchField` reusable to search for products
- Make `SearchField` component recieve two props, `actionWithProductId`
prop to pass the fanctionality that will apply to the search result
whome clicked. `itemIcon` prop to display an icon on the end of the
list items.

- Make the `onClick` prop of the list items of `SearchResultRenderer`
and the icons on the end of the list items passed as props to
`SearchResultRenderer`.

- Pass the nessesary props to `SearchField` in `ProductsManagement` page
after the update of `SearchField` component.

- Move `ItemDisplayer` component from `ProductsDisplayer` component to
new separated component called `DisplayerItemWithLoadingState` and
make it reuseable from anywhere.

- Move `deleteProduct` function and the query of `fetchProduct` function
from `ProductsDisplayer` component to `useProductsActions` hook.

- Rename `formData` variable to `theForm` and `validForm` to `formData`
in `handleSubmit` function in `useAddProducts` hook.

- Create new method called `updateLoading` to update loading state
in `bySteps` function in `useNotifications` hook.

- Add a condition into `isAnImage` function to match url links.

- Rename `h` and `w` properties in `LoadingGrayBar` component to
names have meaning, `h` become `height` and `w` become `width`.

- Use the right props of `LoadingGrayBar` component after changes
in `ProductImagesDisplayer`.