Create edit products page in `products/edit-product`
- Create edit products page in `products/edit-product` path to search
for products using `SearchField` for edit them.

- Create edit products form in `products/edit-product/[productId]` path
so that use the dynamic segment of the path as product's id and fetch
the data of the product and edit it and save the changes.

- Create `useUpdateProduct` hook to handle updating process in there.

- Create `findTheChanges` function in `findUpdateFormChanges` file
to use it for find the changes that made on the product in the form
and return the changes as object.

- Add `edit-product` path under `products` path in `SideBarLinks`.

- Add some conditions to `uploadImages` function in the hook
`useProductImagesUploader`, they are:
"return undefined if the given array of files undefined",
"upload the image only if the type of the item in the array is File",
"push the item directly to `uploadedImagesList` if its type is string".

- Add function to `onClick` prop of edit buttom in `ProductsTableFooter`
for navigate to edit product page by pass the id of the selected product
to `products/edit-product/[productId]` path.

- Add `disabled` and `defaultValue` props to all component that have
`FieldProps` type.

- Use `DisplayerItemWithLoadingState` on `ImageInput` component after
renamed its name from `AddImageInput` to `ImageInput`.

- Replace `Image` component with `Avatar` component in `ImageInput`

- Create `readyUrl` state to deal with `defaultValue` that comes as url.

- Add condition to set the type of `TextField` component to 'text'
in `ImageInput` if there is default value due to the default value is
string (url) and type of `TextField` must be 'text' to can access its
value (the url).