Build sales statistics page and put it in `/` path Create "pages" folder and put `SalesStatisticsPage` in it.
- Delete `moment` library due to we can replace its functionality that we need with
two arrays of monthes names, an array for full names and an array for short names,
these arrays created in CONSTANTS folder at MONTHES file.

- Add return type to `useMonthlyStatistics` hook named `MonthlyStatistics`
which is interface inherits `PromiseState` interface and has one additional prop
which is `monthesData` and the prop is an array of the interface `MonthStatistics`
that created in the same hook file.

- Remove `moreProps` prop in params of `useGetApi` hook.

- Create `pageSpaces` file in `CONSTANTS` folder for the variable that was exist
in `SalesStatisticsPage` page that specify the spaces between elements in app pages
based on screen device size.