Create `CategoriesMonthlySales` component `CategoriesMonthlySales` created for showing how many products sold under each category.
- Replace `TopCategories` component with the new component `CategoriesMonthlySales`
in `ProductsStatisticsPage` page. And call `useBreakPoints` hook to handle
changing the type of `DisplayInfoBox` in the page besed on the screen size.

- Remove calling `useMonthlyCategoriesStatistics` hook in `useProductsStatisticsPageContent` hook,
And also remove `topCategoriesData` variable because the component that supposed
to use it has been deleted.

- Remove the props from `CategoriesMonthlyEarnings` component,
Alternatively Make the component calls `useMonthlyCategoriesStatistics` hook
on its own instead of recieve it as props.
Also remove the title on the y axis of the chart and add custom format to
the lables of y axis.

- Remove `data` prop from return type of `useMonthlyCategoriesStatistics` hook,
And add a new array of datas to `chartData` prop for the new component
`CategoriesMonthlySales` uses it.

- Delete remains of last changes in `CategoriesEarningsPercentages`,
And add `Serieses` section to categories cards.

- Add more option to `useBreakPoints` hook instead of leve it with one option.

- Add more colors to `randomColorsArr` array.

- Use `Paper` component in`ApexchartsContainer` component instead of `Box` component.