Fix naming issue env variables in `UPLOAD_IMAGE_API` Add 'VITE_' in start of any variables to be readable to `Vite.js`
- Add `isUnexpected` and `isUnauthorized` state to `useAdminLogIn` hook
and add them to `App` component as others expected cases while running
the application. And add `logInPage` state in `App` for displaying `LogInForm` component.

- Add warning cases to `handleSubmit` function in `useUpdateProduct` hook,
The first one states if there is no changes made to the product, show a notification
to the user to tell him that there is no changes. The second states if changes happend
and product's images changed but uploading the new images failed fo any reason,
show an error message tells the user that there is one or more images failed to upload.

- Fix types errors in `findUpdateFormChanges` file.

- Remove `isLoading` state of `useMonthlySalesStatistics` hook from `useOrdersPageContent` hook.