feat: split `SettingBox` component to three components
Move the input field in `SettingBox` component to a new components called
`NumberSettingBox` for separating the logic of input field and making it reusable
by other settings components, And move actions buttons section to the new
component `SettingActionsButtons` for reusing it in others setting components,
And let `SettingBox` component work only as container displays the titles and
the descriptions of settings components and receives the input field and actions
buttons section as props.

Make `useMonthlySalesStatistics` hook provides a method for updating the
the monthly sales statistics (`updateMonthlySalesStatistics`).

Inclue `useStatisticsActions` hook which is the hook that supposed to be
included with the last commit, it is the hook that provides the method that
updates the target of year monthes.