feat: Add categories filter to products table in products management page
Create `ProductsTableCategoriesFilter` component which is the input field
for select the categories that based on them the products will filtred abd
put it inside `ProducsTableToolbar` component.

Create `ProductsCategoriesFilterProvider` context provider component and
make it wraps `ProductsManagementPage` component for providing the array
of categories that the products filtred based on them for all nested components
and for avoiding props drilling problem.

Refactor `useProductsPagination` hook and make it uses categories filter array
to include them with pagination requests to the server, And move the function
that sends the HTTP request for paginating the products to `useProductsActions`
hook.

Replace the default pagination component (`GridFooter`) of MUI library with
a customized components inside `ProducsTableFooter` component.