refactor(edit products form): shorten two lines & swap two values
Shorten the lines 21 and 22 in `useProductImagesUploader` hook and
swap the places of `data.url` and `data.secure_url` for giving the priority to the
secure url of the uploaded image to be used.

Make loading state of updating products starts before starting uploading the new
images of the product.

Change the icon of edit products page in `SideBarLinks` component.